integrated standard information into reusable InformationComponent. WCAGComponent no longer needed, deleted.
