+++ src/Information/index.scss
@@ -63,6 +63,12 @@
             font-family: fRegular;
         }
 
+        span {
+            cursor: pointer;
+            display: inline-block;
+            border-bottom: .5px solid $cWhite;
+        }
+
         .wrapperSelector {
             width: 300px;
             height: 10px;
