--- src/WCAGInfo/index.scss
@@ -1,95 +1,0 @@
-.WCAGInfoComponent {
-    position: fixed;
-    width: 100vw;
-    height: 100vh;
-    top: 0;
-    left: 0;
-    background: rgba(51, 51, 51, .9);
-    display: grid;
-    grid-template-columns: 1fr 560px 1fr;
-    grid-template-rows: 30px 48px 94px 470px 1fr;
-    grid-template-areas:
-    '. . . '
-    '.  ButtonComponent  .'
-    '. . . '
-    '.  about  .'
-    '. . .';
-
-    .ButtonComponent {
-        grid-area: ButtonComponent;
-        box-shadow: none;
-        background: $cWhite;
-        color: $cGrey;
-        height: 48px !important;
-        width: 140px;
-        border-radius: 4px;
-        justify-self: center;
-        font-size: $sFont;
-        font-family: fRegular;
-        text-decoration: underline;
-
-        &:hover {
-            animation: none;
-        }
-
-        &:active {
-            box-shadow: none;
-            background: $cWhite;
-        }
-
-        &:focus {
-            border: 0;
-        }
-
-    }// ButtonComponent
-
-    .about {
-        grid-area: about;
-        color: $cWhite;
-        align-self: center;
-
-        div {
-            font-size: 40px;
-            text-align: center;
-            padding-bottom: 50px;
-            font-family: fBold;
-        }
-
-        p {
-            padding-bottom: 30px;
-            line-height: 28px;
-            font-size: 18px;
-            text-align: center;
-            font-family: fRegular;
-        }
-
-        span {
-            cursor: pointer;
-            display: inline-block;
-            border-bottom: .5px solid $cWhite;
-        }
-    }
-
-    @media #{$phone} {
-        grid-template-columns: 20px 1fr 20px;
-    }
-
-    @media #{$tablet} {
-
-        .about {
-            margin-bottom: 80px;
-        }
-    }
-
-    @media screen and (max-width: 320px) {
-
-        .about {
-            div {
-                font-size: 30px;
-                padding-bottom: 40px;
-                padding-top: 50px;
-            }
-        }
-    }
-
-} // WCAGInfoComponent
