+++ src/index.scss
@@ -38,9 +38,7 @@ $tablet: (min-width: 531px), (-webkit-min-device-pixel-ratio: 2) and (min-width:
 @import './PickerIntro/index.scss';
 @import './Picker/index.scss';
 @import './Picker/Selector/index.scss';
-@import './WCAGInfo/index.scss';
 @import './Information/index.scss';
-@import './LightInfo/index.scss';
 @import './Header/index.scss';
 @import './Loading/index.scss';
 @import './Schemes/index.scss';
